Changes wolfram into the individual courses due to removed section
Due to the removal of the main HPC Page, the files on there have been split apart.
Note: The Parallel Computing has GPU Computing as a sub-section. GPU could be removed if it is seen as redundant. 
Another note, if you go to https://reference.wolfram.com/language there are a lot more links. 
Although, much of this appears to be relating to the wolfram language, which may or may not be relevant to a Facilitator's learning goals.