Defines what ROI is
I didn't know what it was at first. This appears to be the first instance (other than the table of contents), so this might be the best place to put a definition of the acronym? 
Maybe it will be okay to leave the other instances (so I am going to leave them for now), but it might be worth looking into removing the others or removing this instance of ROI.